commands: Re-structure command hierarchy
To better align with how other tools (oc, kubectl, eksctl) structure
their commands, and to reduce friction in user adoption, we change how
commands are sent to moactl.

Instead of the original mode of `moactl RESOURCE ACTION`, we invert it
so that it ends up being `moactl ACTION RESOURCE`. The result is a list
of the following commands

    moactl [create|delete|describe|list] [cluster|idp|user]

Other commands remain as before:

    moactl [init|login|logout|version]