logs: Add command to output cluster logs
This new command allows for viewing the cluster Hive installation logs.
By default it shows the last 2000 log entries, which should be
sufficient for the entire installation log on a healthy cluster.
A --tail flag can be passed to view a different number of log lines.

If using the --watch flag, it will poll the hive logs endpoint every few
seconds, appending the last new log entries. It will timeout after an
hour, which should be sufficient time for most clusters to go from
creation to ready.