cluster: Allow the use of --cluster as identifier
When fetching a cluster resource, in addition to passing the cluster
name/ID as a command-line argument, we should also allow it to be sent
as a flag. This allows for more consistent behavior across commands and
resources, while still allowing the current behavior.

With this change, the following commands are equivalent:
* moactl describe cluster foo
* moactl describe cluster --cluster=foo
* moactl describe cluster -c foo

Same rules apply to 'delete' and 'edit'.