addons: Add list and create commands for add-ons
When listing add-ons for a cluster, we show the list of enabled add-ons
that the current organization has quota for, as well as those already
installed in the cluster. From this list the user can use the ID to
manually install add-ons using the `create` command.

Since there can be confusion regarding the spelling of add-ons, I've
added aliases for "addon", "add-on", as well as their plurals. Also,
since the user is not "creating" and add-on, we can use the alias "add"
instead of "create", although both work similarly.