addons: Add list and describe commands for add-ons
The list command used to list add-on installations on a cluster. Now
skipping the --cluster parameter, it lists all add-ons in the system.
This can be used to get the add-on ID for any add-on and then be able to
run the describe command, which will print out a useful description and
documentation for add-ons.