create-cluster: List regions using user AWS creds
To present an accurate list of what regions are supported by the
customer's AWS account, we query the new available_regions API endpoint.
We further filter that list based on the selection of multiple
availability zones: if the user selected multi-az for their cluster, we
narrow down the list to those regions that support it.