verify-oc: Do not error out on invalid version
Detecting an invalid or unsupported version of the openshift-clients is
not an error, but the successful output of the 'verify' command. Since
we also call this command from the 'download' command, we need to ensure
that instead of exiting with an error, the command just ends and allows
the rest of 'download' to go through.