list-versions: Add command to list enabled versions
When creating a cluster, the user is able to pass a parameter with the
version of OpenShift to use. Unless the user goes through the
interactive flow, they won't be able to see which versions are
available. This command allows a user to view a list of all enabled
versions in the current environment.