htpasswd: Add admin resource to login to cluster
To allow customers to login to a cluster without needing to setup an IdP
with a third-party provider and without using the web console, we allow
them to create a cluster-admin user using the htpasswd identity
provider. This auto-generates a password and a cluster-admin user and
then outputs the oc login command necessary to login to the cluster.