idp: Always use interactive mode on unset required flags
For any required attributes, if the user has not set it as a flag in the
CLI, we fallback to displaying the interactive mode for it instead of
failing the request. This also means that for any attribute that the
user has explicitly set through a flag, the tool will not show it even
if the user requested interactive mode.