init: Use explicit login flag checks
When running init, we also ensure the user is logged in. If the user
sets any login flags, we force those as part of the login process,
otherwise we reuse the existing OCM session. By comparing against an
explicit list, we make sure that another flag (e.g. --debug) won't
trigger a re-login with incomplete flags.