Add support for STS clusters
This adds support for passing an AWS Role ARN to be used as
authentication for creating the cluster. Also, it accepts a list of
operator IAM roles so that the OpenShift installer can assume role in
the various components.

Finally, it allows custom tags to be added to all AWS resources created
by the installer.