aws: Filter clusters by AWS account ID
To determine what clusters a user should be able to see, we no longer
limit the filtering to the creator ARN. Instead, we determine what AWS
account was the cluster installed in based on the creator ARN and the
role ARN (for STS clusters). This allows OCM organizations to see and
manage clusters created by different users within the same OCM
organization and same AWS account.