create: Add new account-roles resource
To create STS clusters, users need to have certain roles already created
in their AWS account so that the cluster manager and installer can
authenticate against it. This command creates the roles along with the
necessary policies attached to them.

Also, the operators needed for a successful managed OpenShift cluster
need specific policies that are generic for all clusters of the same
version. These are also created as part of this new command.