create: Add operator-roles command
To create all roles needed by the various OpenShift operators, this new
command loads the pre-configured list of operator IAM roles from the
cluster spec and creates them on the user account with the correct
policies. It also allows a manual mode where the user is given the list
of commands to run.