interactive: Provide real-time validators
Instead of validating user input after the interactive prompt has
returned, we allow a list of validator functions to be passed to the
interactive prompt. This means that the user will get validation
feedback and the program will not exit on invalid data, allowing the
user to correct the input.