sts: Group account roles by prefix
When auto-selecting account roles to use on a cluster, ensure that all
roles match the same prefix group. This ensures consistency and avoids
the auto-find to fall into a mix-and-match approach.

Also extract the prefix from said roles and feed it to the 'create
operator-roles' command so that the correct policies are attached to
those as well.