Modify function `handleErr` to include the error type in the new error
In some places in our code, we display an error message to the user depending
on the error type.
Function `handleErr` omits the error type in the new error it generates.
This function is called in many places in our code, and this change might affect the
error message returned to the user.