Fix bug - deletion of a role with the wrong account ID in role ARN
To delete a role using the AWS client the role name is sufficient.
For `delete ocm-role` and `delete user-role`, we ask the user to enter the whole role ARN,
to be consistent with other ROSA commands.
This PR adds validation to check if the user account ID match the role ARN account ID,
if not, an error message will be displayed and the program will terminate.

Releated: SDA-5544