fedramp: Add environment-specific configuration
To support using ROSA against the FedRAMP deployment of the OCM API, we
override all authentication configuration and API endpoints. We can
determine that the user intends on using FedRAMP if they explicitly set
a GovCloud region, use an encrypted refresh token, or explicitly set the
'govcloud' flag during login.

To allow users to login against the development FedRAMP GovCloud
environment, we add a way of tracking the various APIs and AWS Cognito
user pool and client data.