Fix bug - fetch the subnets from the cluster region
When using the `runtime` object, we create an AWS client with the region
that the user configured locally.
The AWS client fetches only the subnets from that region.
If the cluster is installed in a different region, we'll get
an empty list of subnets.

This PR creates the AWS client with the cluster's region.

Related: SDA-6355