Allow editing default machine pool labels
Related: SDA-7375

Also fixed some issues with interactive mode, in the additional machine
pool. Before the fix, if the user set only labels or taints they we
still asked to set replicas, after the fix the user can set labels,
taints (not for default machinepool), or replicas/autoscaling params.

If the user set at least one of those they will not be prompted for
anything more.
If the user didn't set any of those, they will be prompted for all