fix: checking Account roles against proper Openshift version for the 'create service' command.
Previously, the 'create service' command always checked the Account role against the
highest possible Openshift version. However, this behavior was incorrect as the actual
version used by the service might not be the highest possible one.

To address this issue, the commit modifies the 'create service' command to call the
Managed Service's `version_inquiry` endpoint, which returns the actual version of the
service. This version is then used to check the Account role, ensuring that the role
matches the version that will actually be used.

This change improves the accuracy of the Account role check and prevents potential
issues that could arise from using the wrong version.