refactor: removing init and global variables for hibernate/resume
Global variables can make it difficult to reason about the behavior of
the program since any function can modify them, leading to unexpected
interactions between different parts of the code. By encapsulating the
state in a function, you can better control its access and enforce a
clear boundary around the data.

Similarly, using init() to initialize global state can lead to problems
with testing and debugging since it is executed before the main function
and cannot be easily controlled. By using a function that main calls to
initialize the state, you can more easily control the order of
initialization and ensure that it only occurs when needed.

Overall, using a function that main calls to initialize the state of a
Go program can lead to more maintainable, testable, and reusable code,
making it easier to reason about and improve over time. This commit
starts to position us towards acheiving this by presenting a way to
replace existing global Cmd variables with a function call.