OCM-953 | feat: add automatic upgrades for hypershift control plane
This change adds 2 additional parameters to pass a schedule to the control plane
upgrades.
It also improve the behaviour of interactive mode to be more useful.
Testing has been added by mocking the calls to the backend.
A new dependency is added to check the cron syntax with the same
library used in the backend.
All the changes applies for the moment only to Hosted Control Planes