OCM-16 | feat: ability to set machine pool root disk size
A new CLI flag has been added to the `rosa create cluster` command:

```
--worker-disk-size
```

which sets the root disk size of the worker' nodes.
The new flag accepts Giga/Gigibytes or Tera/Tebi.

Example usage:

```
rosa cluster create --worker-disk-size 128GiB
```

If a giga/tera value is passed it will be converted to Gigibytes, which
is the expected unit format by the AWS API.

The default proposed value remains the same (300 GiB), that the default
flavor uses. This value is fetched from the OCM API endpoint.
The new flag supports both interactive and non-interactive modes.

This resolves: https://issues.redhat.com/browse/OCM-16
This patch is part of this epic
https://issues.redhat.com/browse/SDE-2692

Signed-off-by: Sébastien Han <seb@redhat.com>