OCM-3331 | fix: default values for boolean fields of autoscaler
When using the new autoscaler params, some of them have a default of
true. It makes it impossible to distinguish the two states when
providing the arguments via CLI (in oppose to the interactive mode).

For example, when running rosa create cluster
--autoscaler-balance-similar-node-groups or just rosa create cluster the
same result happens, where balance-similar-node-groups is set to true
for the two cases.