OCM-3911 | feat: use HCP specific endpoints for versions and upgrades
We need to handle HCP versions in the backend side and they may also depend on
the customer.

To obtain this, the current change:
1) introduces the use of an HCP specific version endpoint
2) switch all the available upgrades calls to what is available in the cluster
or the node pool API response, as this is the source of truth.
This will ensure that the list is the correct one and reduce the number of calls in the
client needed to get the version properties.

Functionally there should be no change.

Improved also the testing to use always SDK objects instead of raw JSON which will help
test clarity and mantainence.

Related: OCM-3911