OCM-5530 | feat: add new parameter to disable CNI at cluster creation for HCP
In Hosted Control Planes, we can disable CNI at creation to let user bring their own.

This change exposes the relevant parameter with a new boolean flag.

There is an existing --network-type parameter, but decision was taken not to
reuse this as it is going to be deprecated moving forward and the dedicated parameter
is clearer in terms of usage.

Additional validation added in order to ensure the 2 parameters don't collide and are
correctly validated.

No breaking change expected compared to the current behavior.

Related: OCM-5530