cmd/create/cluster: add a test for the help-text output
We're about to do some refactoring for the way flags are bound and
consumed, so we need to be certain that none of the flag names or their
default values have changed during that process. Since `--help` prints
all of this information in an easily-digestible way, we can test that
the output has not changed during our refactor.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>