cmd/create/cluster: allow user-driven defaults for flags
When we're factoring the codebase to allow programmatic re-use of the
Options concept between different commands, across projects and in
tests, we need to be considerate of the fact that different consumers
will have different opinions on defaults. Furthermore, consumers may
want to operate over options only, instead of messing around with
programmatically setting flag values. Allowing them to pass in their
data directly makes that easy.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>