*: don't assume a *cobra.Command in library code
If multiple callers are to re-use the functionality of a CLI command,
they need to be able to programmatically configure it and run the work
the command would have done, all the while perhaps not even exposing
this in a CLI, or with a *cobra.Command. Factoring helpers to be
ambivalent about where the data comes from and not requiring a
*cobra.Command makes that much easier. A more involved approach here
would have us refactoring all of these helpers to not assume a flag set,
either, but that's a good follow-up for another day.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>