cmd/create/cluster: use normal control flow
Instead of having global variables that hold our state, create structs
and pass them around as necessary. Similarly, return errors when they
occur and allow the top-level caller the ability to determine what to do
with those errors. A top-level caller in a CLI may print to the console
and use os.Exit(). A controller would not. Using os.Exit deep inside of
helpers and library code makes it impossible to re-use that code for
anything else, or run unit tests on it, for that matter.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>