OCM-6484 | fix: generates a valid operator role prefix from the cluster name
A valid operator role prefix has at most 32 characters, but the cluster name can have up to 54 characters this was causing errors as the one reported in https://issues.redhat.com/browse/OCM-6484.
To fix this issue, we use only the first 27 characters of the name if the name is longer than 27 characters.