OCM-7961 | chore: update the fips flag description
This was added as a way to let users know before attempting a deployment, that FIPS
is not available for hosted control plane at this time.  This can be removed at a
time when hosted control plane is FIPS validated.  See
https://docs.aws.amazon.com/rosa/latest/userguide/rosa-deployment-options.html#rosa-deployment-model-differences
for more details.

Signed-off-by: Dustin Scott <dscott@redhat.com>

OCM-7961 | chore: add space to split sentence

Signed-off-by: Dustin Scott <dscott@redhat.com>