OCM-8337 | fix: [HCP] always set replicas when updating machinepools
We always pass the replicas when editing a machinepool, even if
the value has not changed, as this will then be a no-op. Right now
we are not setting replicas if they are the same as the current value
and this result in an error returned to the user.

Related: OCM-8337